fix: bug in status of the milestones
